split generic code from OSMScout2 demo to libosmscout-client-qt library, setup CMake to build it
