+++ libosmscout-client-qt/include/osmscout/DBThread.h
