+++ libosmscout-client-qt/include/osmscout/MapWidget.h
