+++ libosmscout-client-qt/include/osmscout/RoutingModel.h
