+++ libosmscout-client-qt/include/osmscout/SearchLocationModel.h
