+++ libosmscout-client-qt/include/osmscout/Settings.h
