+++ libosmscout-client-qt/src/DBThread.cpp
@@ -17,7 +17,7 @@
  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "DBThread.h"
+#include "osmscout/DBThread.h"
 
 #include <iostream>
