+++ libosmscout-client-qt/src/MapWidget.cpp
@@ -17,7 +17,7 @@
  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "MapWidget.h"
+#include "osmscout/MapWidget.h"
 
 #include <iostream>
