+++ libosmscout-client-qt/src/RoutingModel.cpp
@@ -17,7 +17,7 @@
  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "RoutingModel.h"
+#include "osmscout/RoutingModel.h"
 
 #include <iostream>
 #include <iomanip>
